'3.7.3'

Manually resolved conflicts:
	SConstruct
	buildscripts/setup_multiversion_mongodb.py
	buildscripts/smoke.py (this file was deleted)
	src/mongo/db/server_options_helpers.cpp

+++ jstests/concurrency/fsm_workloads/invalidated_cursors.js
@@ -7,6 +7,9 @@
  * cursor manager. Threads perform find, getMore and explain commands while the database,
  * collection, or an index is dropped.
  */
+
+load('jstests/concurrency/fsm_workload_helpers/server_types.js');  // for isMongos
+
 var $config = (function() {
 
     let data = {
@@ -41,6 +44,31 @@ var $config = (function() {
                 assertAlways.commandWorked(db[collName].createIndex(indexSpec));
             });
         },
+
+        /**
+         * Calls 'killFn' on a random getMore that's currently running.
+         */
+        killRandomGetMore: function killRandomGetMore(someDB, killFn) {
+            const admin = someDB.getSiblingDB("admin");
+            const getMores = admin
+                                 .aggregate([
+                                     // idleConnections true so we can also kill cursors which are
+                                     // not currently active.
+                                     {$currentOp: {idleConnections: true}},
+                                     // We only about getMores.
+                                     {$match: {"command.getMore": {$exists: true}}},
+                                     // Only find getMores running on the database for this test.
+                                     {$match: {"ns": this.uniqueDBName + ".$cmd"}}
+                                 ])
+                                 .toArray();
+
+            if (getMores.length === 0) {
+                return;
+            }
+
+            const toKill = this.chooseRandomlyFrom(getMores);
+            return killFn(toKill);
+        }
     };
 
     let states = {
@@ -53,8 +81,8 @@ var $config = (function() {
          * Runs a query on the collection with a small enough batchSize to leave the cursor open.
          * If the command was successful, stores the resulting cursor in 'this.cursor'.
          */
-        query: function query(db, collName) {
-            let myDB = db.getSiblingDB(this.uniqueDBName);
+        query: function query(unusedDB, unusedCollName) {
+            let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
             let res = myDB.runCommand({
                 find: this.chooseRandomlyFrom(this.involvedCollections),
                 filter: {},
@@ -62,15 +90,15 @@ var $config = (function() {
             });
 
             if (res.ok) {
-                this.cursor = new DBCommandCursor(db, res, this.batchSize);
+                this.cursor = new DBCommandCursor(myDB, res, this.batchSize);
             }
         },
 
         /**
          * Explains a find on a collection.
          */
-        explain: function explain(db, collName) {
-            let myDB = db.getSiblingDB(this.uniqueDBName);
+        explain: function explain(unusedDB, unusedCollName) {
+            let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
             let res = myDB.runCommand({
                 explain: {find: this.chooseRandomlyFrom(this.involvedCollections), filter: {}},
                 verbosity: "executionStats"
@@ -78,17 +106,51 @@ var $config = (function() {
             assertAlways.commandWorked(res);
         },
 
-        killCursor: function killCursor(db, collName) {
-            if (this.hasOwnProperty('cursor')) {
-                this.cursor.close();
+        /**
+         * This is just a transition state that serves as a placeholder to delegate to one of the
+         * specific kill types like 'killOp' or 'killCursors'.
+         */
+        kill: function kill(unusedDB, unusedCollName) {},
+
+        /**
+         * Choose a random cursor that's open and kill it.
+         */
+        killCursor: function killCursor(unusedDB, unusedCollName) {
+            if (isMongos(unusedDB)) {
+                // SERVER-33700: We can't list operations running locally on a mongos.
+                return;
             }
+
+            const myDB = unusedDB.getSiblingDB(this.uniqueDBName);
+
+            // Not checking the return value, since the cursor may be closed on its own
+            // before this has a chance to run.
+            this.killRandomGetMore(myDB, function(toKill) {
+                const res = myDB.runCommand(
+                    {killCursors: toKill.command.collection, cursors: [toKill.command.getMore]});
+                assertAlways.commandWorked(res);
+            });
+        },
+
+        killOp: function killOp(unusedDB, unusedCollName) {
+            if (isMongos(unusedDB)) {
+                // SERVER-33700: We can't list operations running locally on a mongos.
+                return;
+            }
+
+            const myDB = unusedDB.getSiblingDB(this.uniqueDBName);
+            // Not checking return value since the operation may end on its own before we have
+            // a chance to kill it.
+            this.killRandomGetMore(myDB, function(toKill) {
+                assertAlways.commandWorked(myDB.killOp(toKill.opid));
+            });
         },
 
         /**
          * Requests enough results from 'this.cursor' to ensure that another batch is needed, and
          * thus ensures that a getMore request is sent for 'this.cursor'.
          */
-        getMore: function getMore(db, collName) {
+        getMore: function getMore(unusedDB, unusedCollName) {
             if (!this.hasOwnProperty('cursor')) {
                 return;
             }
@@ -101,12 +163,15 @@ var $config = (function() {
                     this.cursor.next();
                 } catch (e) {
                     // The getMore request can fail if the database, a collection, or an index was
-                    // dropped.
+                    // dropped. It can also fail if another thread kills it through killCursor or
+                    // killOp.
                     assertAlways.contains(e.code,
                                           [
                                             ErrorCodes.OperationFailed,
                                             ErrorCodes.QueryPlanKilled,
-                                            ErrorCodes.CursorNotFound
+                                            ErrorCodes.CursorNotFound,
+                                            ErrorCodes.CursorKilled,
+                                            ErrorCodes.Interrupted,
                                           ],
                                           'unexpected error code: ' + e.code + ': ' + e.message);
                 }
@@ -117,8 +182,15 @@ var $config = (function() {
          * Drops the database being used by this workload and then re-creates each of
          * 'this.involvedCollections' by repopulating them with data and indexes.
          */
-        dropDatabase: function dropDatabase(db, collName) {
-            let myDB = db.getSiblingDB(this.uniqueDBName);
+        dropDatabase: function dropDatabase(unusedDB, unusedCollName) {
+            if (isMongos(unusedDB)) {
+                // SERVER-17397: Drops in a sharded cluster may not fully succeed. It is not safe
+                // to drop and then recreate a collection with the same name, so we skip dropping
+                // and recreating the database.
+                return;
+            }
+
+            let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
             myDB.dropDatabase();
 
             // Re-create all of the collections and indexes that were dropped.
@@ -131,8 +203,14 @@ var $config = (function() {
          * Randomly selects a collection from 'this.involvedCollections' and drops it. The
          * collection is then re-created with data and indexes.
          */
-        dropCollection: function dropCollection(db, collName) {
-            let myDB = db.getSiblingDB(this.uniqueDBName);
+        dropCollection: function dropCollection(unusedDB, unusedCollName) {
+            if (isMongos(unusedDB)) {
+                // SERVER-17397: Drops in a sharded cluster may not fully succeed. It is not safe
+                // to drop and then recreate a collection with the same name, so we skip it.
+                return;
+            }
+
+            let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
             let targetColl = this.chooseRandomlyFrom(this.involvedCollections);
 
             myDB[targetColl].drop();
@@ -146,8 +224,8 @@ var $config = (function() {
          * 'this.indexSpecs' and drops that particular index from the collection. The index is then
          * re-created.
          */
-        dropIndex: function dropIndex(db, collName) {
-            let myDB = db.getSiblingDB(this.uniqueDBName);
+        dropIndex: function dropIndex(unusedDB, unusedCollName) {
+            let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
             let targetColl = this.chooseRandomlyFrom(this.involvedCollections);
             let indexSpec = this.chooseRandomlyFrom(this.indexSpecs);
 
@@ -169,29 +247,31 @@ var $config = (function() {
             dropIndex: 0.1,
         },
 
-        query: {killCursor: 0.1, getMore: 0.9},
+        query: {kill: 0.1, getMore: 0.9},
         explain: {explain: 0.1, init: 0.9},
+        kill: {killOp: 0.5, killCursor: 0.5},
+        killOp: {init: 1},
         killCursor: {init: 1},
-        getMore: {killCursor: 0.2, getMore: 0.6, init: 0.2},
+        getMore: {kill: 0.2, getMore: 0.6, init: 0.2},
         dropDatabase: {init: 1},
         dropCollection: {init: 1},
         dropIndex: {init: 1}
     };
 
-    function setup(db, collName, cluster) {
+    function setup(unusedDB, unusedCollName, cluster) {
         // Use the workload name as part of the database name, since the workload name is assumed to
         // be unique.
-        this.uniqueDBName = db.getName() + 'invalidated_cursors';
+        this.uniqueDBName = unusedDB.getName() + 'invalidated_cursors';
 
-        let myDB = db.getSiblingDB(this.uniqueDBName);
+        let myDB = unusedDB.getSiblingDB(this.uniqueDBName);
         this.involvedCollections.forEach(collName => {
             this.populateDataAndIndexes(myDB, collName);
             assertAlways.eq(this.numDocs, myDB[collName].find({}).itcount());
         });
     }
 
-    function teardown(db, collName, cluster) {
-        db.getSiblingDB(this.uniqueDBName).dropDatabase();
+    function teardown(unusedDB, unusedCollName, cluster) {
+        unusedDB.getSiblingDB(this.uniqueDBName).dropDatabase();
     }
 
     return {
